Source code/requirements.txt
glcontext==2.5.0
moderngl==5.10.0
numpy==1.26.4
pygame==2.5.2
PyGLM==2.7.1
PyOpenGL==3.1.7
PyOpenGL-accelerate==3.1.7
PyWave==0.5.0
PyWavefront==1.3.3